requirements.txt
asgiref==3.2.5
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.4
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.3
pylint==2.4.4
pyparsing==2.4.6
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0